backend/requirements.txt
fastapi
uvicorn
supabase
langchain
psycopg2-binary
python-dotenv
pypdf
tiktoken
sentence-transformers
langchain-google-genai
langchain-community
python-multipart
